+++ modules/error_manager/builtin/include/getLastReportBuiltin.hpp
@@ -23,30 +23,17 @@
 // License along with this program. If not, see <http://www.gnu.org/licenses/>.
 // LICENCE_BLOCK_END
 //=============================================================================
-#include "NelsonGateway.hpp"
-#include "errorBuiltin.hpp"
-#include "lasterrorBuiltin.hpp"
-#include "lastwarnBuiltin.hpp"
-#include "warningBuiltin.hpp"
-#include "getLastReportBuiltin.hpp"
+#pragma once
 //=============================================================================
-using namespace Nelson;
+#include "ArrayOf.hpp"
+#include "Evaluator.hpp"
 //=============================================================================
-const std::wstring gatewayName = L"error_manager";
+namespace Nelson {
 //=============================================================================
-static const nlsGateway gateway[] = {
-    { "error", Nelson::ErrorManagerGateway::errorBuiltin, 0, 1 },
-    { "warning", Nelson::ErrorManagerGateway::warningBuiltin, 1, -1 },
-    { "lasterror", Nelson::ErrorManagerGateway::lasterrorBuiltin, 1, 1 },
-    { "lastwarn", Nelson::ErrorManagerGateway::lastwarnBuiltin, 2, 2 },
-    { "getLastReport", Nelson::ErrorManagerGateway::getLastReportBuiltin, 1, 0 },
-};
+namespace ErrorManagerGateway {
+    ArrayOfVector
+    getLastReportBuiltin(Evaluator* eval, int nLhs, const ArrayOfVector& argIn);
+}
 //=============================================================================
-NLSGATEWAYFUNC(gateway)
-//=============================================================================
-NLSGATEWAYINFO(gateway)
-//=============================================================================
-NLSGATEWAYREMOVE(gateway)
-//=============================================================================
-NLSGATEWAYNAME()
+} // namespace Nelson
 //=============================================================================